requirements.txt
numpy
pandas
scipy
matplotlib
Pillow
torch
torchvision
scikit-learn
pygad
tqdm
seaborn
PyYAML
